+++ packages/vendure-plugin-utm-tracker/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0 (2025-10-24)
+
+- Allow adding bulk UTM parameters to an order
+- Allow passing `connectedAt` date from client, because UTM parameters are saved on the client first
+
 # 1.0.0 (2025-10-23)
 
 - Initial release